+++ cpp/src/error.cpp
@@ -10,7 +10,7 @@
 #include <stdlib.h>
 #include <string>
 
-const error_t* error_new(std::string prog_name, unsigned int line, unsigned int pos, std::string str) {
+const error_t* error_new(std::string prog_name, size_t line, size_t pos, std::string str) {
     error_t* err = (error_t*)calloc(1,sizeof(error_t));
     err->prog_name = prog_name;
     err->line = line;
